requirements.txt
streamlit
matplotlib
numpy
groq
python-dotenv
